device_driver_common::span

Struct Span

pub struct Span {
    pub start: usize,
    pub end: usize,
}

Fields
start: usizeend: usize
Implementations

impl Span

pub const fn empty() -> Self

pub fn is_empty(&self) -> bool

pub fn or(&self, other: Self) -> Self

Return self if not empty, or the other span if self is empty

pub fn to(&self, other: Self) -> Self

Get the combined span that encompasses both spans. The order does not matter.

pub fn skip(&self, skip: Self) -> Self

Take the current span, but skip the start until the skip span has been passed

pub fn start_from(&self, start: Self) -> Self

Take the current span, but skip the first part and start from the given start span

pub fn collapse_to_end(&self) -> Span

Discard the span and keep a zero-length span at the end value